fix(fs): Handle absolute paths for ignore rules file
If the path is an absolute path, we should use it as it is instead of appending
the base directory to it. E.g. we have a use case where use have:

```
Files policy:
  Ignore cache directories:             true   (defined for this target)
  No ignore rules:
  Read ignore rules from files:                (defined for this target)
    .kopiaignore
    /opt/hostbackup/.kopiaignore
  Scan one filesystem only:            false   (defined for this target)
```

And we don't want to use CLI-aware stuff, e.g. `kopia ... --global --add-ignore ...`,
because it's hard to manage the list (adding, removing) idempotent.

Signed-off-by: Donatas Abraitis <donatas@opensourcerouting.org>